+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [2.5.2](https://github.com/bitrix24/b24ui/compare/v2.5.1...v2.5.2) (2026-03-xx)
 
 * fix(NavigationMenu): improve theme
+* fix(DashboardSidebar|Header): improve menu
+* docs(dashboard*): improve
+* playgrounds: improve page shortcuts
 
 ## [2.5.1](https://github.com/bitrix24/b24ui/compare/v2.4.2...v2.5.1) (2026-03-24)